+++ simplesockets/securesockets/PyCryptodome_Socket.py
@@ -1,3 +1,5 @@
+import traceback
+
 from simplesockets._support_files import cipher
 from simplesockets.simple_sockets import TCPClient, TCPServer
 
@@ -61,22 +63,16 @@ class SecureClient(TCPClient):
 
         super().setup(target_ip, target_port, recv_buffer, on_connect, on_disconnect, on_receive)
 
-    def connect(self, user: str, pw: str) -> bool:
+    def login_data(self, user: str, password):
         """
-        function tries to connect
+        overwrites self.user and self.pw with given arguments
 
         Args:
             user: username
-            pw: password
-
-        Returns:
-            returns True if connecting was successful
-
+            password: password
         """
-
         self.user = user
-        self.pw = pw
-        return super().connect()
+        self.pw = password
 
     def on_connect(self) -> bool:
         """
@@ -109,7 +105,8 @@ class SecureClient(TCPClient):
                         keys[user.decode()] = cipher.import_asym_key(key)
                 self.users_keys = keys.copy()
             except Exception as e:
-                print(e)
+                self.event.exception.occurred = True
+                self.event.exception.exceptions.add(e, traceback.format_exc())
             return True
 
     def recv_data(self) -> tuple:
@@ -126,13 +123,18 @@ class SecureClient(TCPClient):
         target, data = rest.split(self.seperators[1])
 
         #Decrypt
-        type = cipher.decrypt_asym(type, self.own_keys[0])
-        target = cipher.decr_data(target, prkey=self.own_keys[0], output="bytes")
+        try:
+            type = cipher.decrypt_asym(type, self.own_keys[0])
+            target = cipher.decr_data(target, prkey=self.own_keys[0], output="bytes")
+        except ValueError:
+            pass
+
         data = cipher.decr_data(data, prkey=self.own_keys[0], output="bytes")
 
         return (target, type, data)
 
-    def send_data(self, target: bytes, type: bytes, data: bytes, username: str = None, key: RsaKey = None) -> bool:
+    def send_data(self, target: bytes, type: bytes, data: bytes, username: Optional[str] = None,
+                  key: Optional[RsaKey] = None, encr_rest: bool = True) -> bool:
         """
         Sends data to the the username or given socket encrypted with a key
 
@@ -151,15 +153,20 @@ class SecureClient(TCPClient):
         if key is not None and username is None:
             pass
         elif key is None and username is not None:
-            key = self.users_keys.get(username)
+            if username is not self.user:
+                key = self.users_keys.get(username)
+            else:
+                key = self.own_keys[1]
         else:
             return False
 
         if key is None:
             return False
 
-        target = cipher.encr_data(target, self.server_key)
-        type = cipher.encrypt_asym(type, self.server_key)
+        if encr_rest:
+            target = cipher.encr_data(target, self.server_key)
+            type = cipher.encrypt_asym(type, self.server_key)
+
         data = cipher.encr_data(data, key)
 
         to_send = type + self.seperators[0] + target + self.seperators[1] + data
@@ -212,9 +219,10 @@ class SecureServer(TCPServer):
 
         self.indent = 4
 
-    def setup(self, filepath: str, ip: Optional[str] = "127.0.0.1", port: Optional[int] = 25567,
-              listen: Optional[int] = 5, recv_buffer: int = 2048, handle_client: Callable = None, on_connect: Callable = None,
-              on_disconnect: Callable = None, on_receive: Callable = None):
+    def setup(self, ip: Optional[str] = "127.0.0.1", port: Optional[int] = 25567,
+              listen: Optional[int] = 5, recv_buffer: Optional[int] = 2048, handle_client: Optional[Callable] = None,
+              on_connect: Optional[Callable] = None, on_disconnect: Optional[Callable] = None,
+              on_receive: Optional[Callable] = None, filepath: str = None):
         """
         prepares the Server
 
@@ -229,16 +237,27 @@ class SecureServer(TCPServer):
             on_disconnect: function that will be executed on disconnection, it takes the address(tuple) as an argument
             on_receive: function that will be executed on receive, it takes the clientsocket, address, received data
 
-
+        Raises:
+            Exception: When filepath is missing and no alternative on_connect function is given
         """
-
         if on_connect is None:
             on_connect = self.on_connect
 
         if on_disconnect is None:
             on_disconnect = self.on_disconnect
 
-        self.filepath = os.path.expanduser(filepath)
+        if isinstance(filepath, str):
+            try:
+                filepath = os.path.expanduser(filepath)
+                with open(filepath, "r") as file:
+                    json.load(file)
+            except Exception:
+                self.filepath = None
+                if on_connect is not None:
+                    raise Exception("You must give a filepath to a json file, containing information about users "
+                                    "and their passwords")
+            else:
+                self.filepath = filepath
 
         super().setup(ip, port, listen, recv_buffer, handle_client, on_connect=on_connect,
                       on_disconnect=on_disconnect, on_receive=on_receive)
@@ -412,14 +431,17 @@ class SecureServer(TCPServer):
         target, data = rest.split(self.seperators[1])
 
         # Decrypt
-        type = cipher.decrypt_asym(type, self.own_keys[0])
-        target = cipher.decr_data(target, prkey=self.own_keys[0], output="bytes")
+        try:
+            type = cipher.decrypt_asym(type, self.own_keys[0])
+            target = cipher.decr_data(target, prkey=self.own_keys[0], output="bytes")
+        except ValueError:
+            pass
 
         return (target, type, data)
 
     def send_data(self, target: bytes, type: bytes, data: bytes, username: Optional[str] = None,
                   key: Optional[RsaKey] = None, client_socket: Optional[socket.socket] = None,
-                  encr_data: Optional[bool] = True) -> bool:
+                  encr_data: Optional[bool] = True, encr_rest: Optional[bool] = True) -> bool:
         """
         function sends encrypted data to a user or socket
 
@@ -442,17 +464,23 @@ class SecureServer(TCPServer):
         elif key is None and client_socket is None and username is not None:
             key = self.client_keys.get(username)
             client_socket = self.clients.get(self.get_address_by_user(username))[1]
+        elif key is None and client_socket is not None and username is None and encr_data is False and encr_rest is False:
+            pass
+        elif encr_rest is False and encr_data is False and client_socket is not None:
+            pass
         else:
             self.event.exception.occurred = True
-            self.event.exception._list.append((Exception("You must give an username or key and clientsocket")))
+            self.event.exception.exceptions.add(Exception("You must give an username or key and clientsocket", traceback.format_exc()))
             return False
 
-        target = cipher.encr_data(target, key)
-        type = cipher.encrypt_asym(type, key)
+        if encr_rest:
+            target = cipher.encr_data(target, key)
+            type = cipher.encrypt_asym(type, key)
         if encr_data:
             data = cipher.encr_data(data, key)
 
-        to_send = type + self.seperators[0] + target + self.seperators[1] + data
+        to_send = b"".join([type, self.seperators[0], target, self.seperators[1], data])
+        #type + self.seperators[0] + target + self.seperators[1] + data
         return super().send_data(to_send, client_socket)
 
     def get_client_keys(self) -> dict: