+++ simplesockets/securesockets/SelfmadeCipher_Socket.py
@@ -8,6 +8,7 @@ import json
 import socket
 import time
 import os
+import traceback
 
 
 class SecureClient(TCPClient):
@@ -25,7 +26,7 @@ class SecureClient(TCPClient):
         """
         super().__init__()
         self.user: str = ""
-        self.pw: str = ""
+        self.pw = ""
 
         self.seperators: list = [b'type_targ_sepLnpEwEljZi', b'targ_data_sepcLkGqydgGY']
 
@@ -58,7 +59,7 @@ class SecureClient(TCPClient):
 
     def setup(self, target_ip: str, target_port: Optional[int] = 25567, recv_buffer: Optional[int] = 2048,
               on_connect: Optional[Callable] = None, on_disconnect: Optional[Callable] = None,
-              on_receive: Optional[Callable]=None):
+              on_receive: Optional[Callable] = None):
         """
         function sets up the Client
 
@@ -72,25 +73,20 @@ class SecureClient(TCPClient):
 
         """
         if on_connect is None:
-            on_connect = self.on_connect
+            on_connect = self._on_connect
 
         super().setup(target_ip, target_port, recv_buffer, on_connect, on_disconnect, on_receive)
 
-    def connect(self, user: str, pw: str) -> bool:
+    def login_data(self, user: str, password):
         """
-        function tries to connect
+        overwrites self.user and self.pw with given arguments
 
         Args:
             user: username
-            pw: password
-
-        Returns:
-            returns True if connecting was successful
-
+            password: password
         """
         self.user = user
-        self.pw = pw
-        return super().connect()
+        self.pw = password
 
     def on_connect(self) -> bool:
         """
@@ -123,7 +119,8 @@ class SecureClient(TCPClient):
                         keys[user.decode()] = RSA.import_key(key)
                 self.users_keys = keys.copy()
             except Exception as e:
-                print(e)
+                self.event.exception.occurred = True
+                self.event.exception.exceptions.add(e, traceback.format_exc())
             return True
 
     def recv_data(self) -> tuple:
@@ -139,14 +136,18 @@ class SecureClient(TCPClient):
         target, data = rest.split(self.seperators[1])
 
         #Decrypt
-        type = self.own_keys[0]._decrypt(type, 1)
-        target = self.__decypt_data(target, self.own_keys[0]).decode()
+        try:
+            type = self.own_keys[0]._decrypt(type, 1)
+            target = self.__decypt_data(target, self.own_keys[0]).decode()
+        except ValueError:
+            pass
+
         data = self.__decypt_data(data, self.own_keys[0])
 
         return (target, type, data)
 
     def send_data(self, target: bytes, type: bytes, data: bytes, username: Optional[str] = None,
-                  key: Optional[RSA.RSA_PUBLIC_KEY] = None) -> bool:
+                  key: Optional[RSA.RSA_PUBLIC_KEY] = None, encr_rest: bool = True) -> bool:
         """
         Sends data to the the username or given socket encrypted with a key
 
@@ -165,15 +166,20 @@ class SecureClient(TCPClient):
         if key is not None and username is None:
             pass
         elif key is None and username is not None:
-            key = self.users_keys.get(username)
+            if username is not self.user:
+                key = self.users_keys.get(username)
+            else:
+                key = self.own_keys[1]
         else:
             return False
 
         if key is None:
             return False
 
-        target = self.__enrcypt_data(target, self.server_key)
-        type = self.server_key.encrypt(type, 1)
+        if encr_rest:
+            target = self.__enrcypt_data(target, self.server_key)
+            type = self.server_key.encrypt(type, 1)
+
         data = self.__enrcypt_data(data, key)
 
         to_send = type + self.seperators[0] + target + self.seperators[1] + data
@@ -238,10 +244,10 @@ class SecureServer(TCPServer):
         pad = b_veginer.import_pad(pad)
         return pad.decrypt(enrcypted)
 
-    def setup(self, filepath: str, ip: Optional[str] = "127.0.0.1", port: Optional[int] = 25567,
+    def setup(self, ip: Optional[str] = "127.0.0.1", port: Optional[int] = 25567,
               listen: Optional[int] = 5, recv_buffer: Optional[int] = 2048, handle_client: Optional[Callable] = None,
               on_connect: Optional[Callable] = None, on_disconnect: Optional[Callable] = None,
-              on_receive: Optional[Callable] = None):
+              on_receive: Optional[Callable] = None, filepath: str = None):
         """
         prepares the Server
 
@@ -256,6 +262,8 @@ class SecureServer(TCPServer):
             on_disconnect: function that will be executed on disconnection, it takes the address(tuple) as an argument
             on_receive: function that will be executed on receive, it takes the clientsocket, address, received data
 
+        Raises:
+            Exception: When filepath is missing and no alternative on_connect function is given
         """
         if on_connect is None:
             on_connect = self.on_connect
@@ -263,7 +271,18 @@ class SecureServer(TCPServer):
         if on_disconnect is None:
             on_disconnect = self.on_disconnect
 
-        self.filepath = os.path.expanduser(filepath)
+        if isinstance(filepath, str):
+            try:
+                filepath = os.path.expanduser(filepath)
+                with open(filepath, "r") as file:
+                    json.load(file)
+            except Exception:
+                self.filepath = None
+                if on_connect is not None:
+                    raise Exception("You must give a filepath to a json file, containing information about users "
+                                    "and their passwords")
+            else:
+                self.filepath = filepath
 
         super().setup(ip, port, listen, recv_buffer, handle_client, on_connect=on_connect,
                       on_disconnect=on_disconnect, on_receive=on_receive)
@@ -309,9 +328,13 @@ class SecureServer(TCPServer):
             address: the address, containing Ip and Port
 
         """
-        username = self.users.pop(address)
-        self.client_keys.pop(username)
-        self.to_send_client_keys.pop(username)
+        try:
+            username = self.users.pop(address)
+            self.client_keys.pop(username)
+            self.to_send_client_keys.pop(username)
+        except Exception as e:
+            self.event.exception.occurred = True
+            self.event.exception.exceptions.add(e, traceback.format_exc())
 
     def load_users(self, get: Optional[str] = None) -> dict:  # json
         """
@@ -427,14 +450,16 @@ class SecureServer(TCPServer):
         target, data = rest.split(self.seperators[1])
 
         # Decrypt
-        type = self.own_keys[0].decrypt(type, 1)
-        target = self.__decypt_data(target, self.own_keys[0])
-
+        try:
+            type = self.own_keys[0].decrypt(type, 1)
+            target = self.__decypt_data(target, self.own_keys[0])
+        except ValueError:
+            pass
         return (target, type, data)
 
     def send_data(self, target: bytes, type: bytes, data: bytes, username: Optional[str] = None,
                   key: Optional[RSA.RSA_PUBLIC_KEY] = None, client_socket: Optional[socket.socket] = None,
-                  encr_data: Optional[bool] = True) -> bool:
+                  encr_data: Optional[bool] = True, encr_rest: bool = True) -> bool:
         """
         function sends encrypted data to a user or socket
 
@@ -454,15 +479,21 @@ class SecureServer(TCPServer):
         if key is not None and client_socket is not None and username is None:
             pass
         elif key is None and client_socket is None and username is not None:
-            key: RSA.RSA_PUBLIC_KEY = self.client_keys.get(username)
+            key = self.client_keys.get(username)
             client_socket = self.clients.get(self.get_address_by_user(username))[1]
+        elif key is None and client_socket is not None and username is None and encr_data is False and encr_rest is False:
+            pass
+        elif encr_rest is False and encr_data is False and client_socket is not None:
+            pass
         else:
             self.event.exception.occurred = True
-            self.event.exception._list.append((Exception("You must give an username or key and clientsocket")))
+            self.event.exception.exceptions.add(
+                Exception("You must give an username or key and clientsocket", traceback.format_exc()))
             return False
 
-        target = self.__enrcypt_data(target, key)
-        type = key.encrypt(type, 1)
+        if encr_rest:
+            target = self.__enrcypt_data(target, key)
+            type = key.encrypt(type, 1)
         if encr_data:
             data = self.__enrcypt_data(data, key)